pyproject.toml
[tool.poetry]
name = "openai-whisper"
version = "0.1.0"
description = ""
authors = ["DOUIF <a15975345678@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
openai = "^1.34.0"
python-dotenv = "^1.0.1"
aiofiles = "^23.2.1"
pydub = "^0.25.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
